+++ tests/Faithlife.Data.Tests/DbConnectorTests.cs
@@ -303,22 +303,22 @@ namespace Faithlife.Data.Tests
 		}
 
 		[Test]
-		public void PrepareTests()
+		public void CacheTests()
 		{
 			using var connector = CreateConnector();
 			connector.Command("create table Items (ItemId integer primary key, Name text not null);").Execute().Should().Be(0);
 			foreach (var name in new[] { "one", "two", "three" })
-				connector.Command("insert into Items (Name) values (@name);", ("name", name)).Prepare().Execute().Should().Be(1);
+				connector.Command("insert into Items (Name) values (@name);", ("name", name)).Cache().Execute().Should().Be(1);
 			connector.Command("select Name from Items order by ItemId;").Query<string>().Should().Equal("one", "two", "three");
 		}
 
 		[Test]
-		public async Task PrepareAsyncTests()
+		public async Task CacheAsyncTests()
 		{
 			await using var connector = CreateConnector();
 			(await connector.Command("create table Items (ItemId integer primary key, Name text not null);").ExecuteAsync()).Should().Be(0);
 			foreach (var name in new[] { "one", "two", "three" })
-				(await connector.Command("insert into Items (Name) values (@name);", ("name", name)).Prepare().ExecuteAsync()).Should().Be(1);
+				(await connector.Command("insert into Items (Name) values (@name);", ("name", name)).Cache().ExecuteAsync()).Should().Be(1);
 			(await connector.Command("select Name from Items order by ItemId;").QueryAsync<string>()).Should().Equal("one", "two", "three");
 		}